+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Parsing of `XlsxRow#style_index`
 * New `Worksheet#rows_raw` field
 * Add parsing `Chart#relationships`
+* Add parsing `Chart#style`
 
 ### Changes